+++ src/RequestGroup.cc
@@ -390,7 +390,7 @@ void RequestGroup::createInitialCommand
       }
       removeDefunctControlFile(progressInfoFile);
       {
-        uint64_t actualFileSize = pieceStorage_->getDiskAdaptor()->size();
+        off_t actualFileSize = pieceStorage_->getDiskAdaptor()->size();
         if(actualFileSize == downloadContext_->getTotalLength()) {
           // First, make DiskAdaptor read-only mode to allow the
           // program to seed file in read-only media.
@@ -555,7 +555,7 @@ void RequestGroup::processCheckIntegrityEntry
  const SharedHandle<CheckIntegrityEntry>& entry,
  DownloadEngine* e)
 {
-  uint64_t actualFileSize = pieceStorage_->getDiskAdaptor()->size();
+  off_t actualFileSize = pieceStorage_->getDiskAdaptor()->size();
   if(actualFileSize > downloadContext_->getTotalLength()) {
     entry->cutTrailingGarbage();
   }
@@ -863,7 +863,7 @@ std::string RequestGroup::getFirstFilePath() const
   }
 }
 
-uint64_t RequestGroup::getTotalLength() const
+off_t RequestGroup::getTotalLength() const
 {
   if(!pieceStorage_) {
     return 0;
@@ -876,7 +876,7 @@ uint64_t RequestGroup::getTotalLength() const
   }
 }
 
-uint64_t RequestGroup::getCompletedLength() const
+off_t RequestGroup::getCompletedLength() const
 {
   if(!pieceStorage_) {
     return 0;
@@ -902,8 +902,8 @@ void RequestGroup::validateFilename(const std::string& expectedFilename,
   }
 }
 
-void RequestGroup::validateTotalLength(uint64_t expectedTotalLength,
-                                       uint64_t actualTotalLength) const
+void RequestGroup::validateTotalLength(off_t expectedTotalLength,
+                                       off_t actualTotalLength) const
 {
   if(expectedTotalLength <= 0) {
     return;
@@ -921,7 +921,7 @@ void RequestGroup::validateFilename(const std::string& actualFilename) const
   validateFilename(downloadContext_->getFileEntries().front()->getBasename(), actualFilename);
 }
 
-void RequestGroup::validateTotalLength(uint64_t actualTotalLength) const
+void RequestGroup::validateTotalLength(off_t actualTotalLength) const
 {
   validateTotalLength(getTotalLength(), actualTotalLength);
 }
@@ -1155,7 +1155,7 @@ void RequestGroup::setProgressInfoFile(const BtProgressInfoFileHandle& progressI
 bool RequestGroup::needsFileAllocation() const
 {
   return isFileAllocationEnabled() &&
-    (uint64_t)option_->getAsLLInt(PREF_NO_FILE_ALLOCATION_LIMIT) <= getTotalLength() &&
+    option_->getAsLLInt(PREF_NO_FILE_ALLOCATION_LIMIT) <= getTotalLength() &&
     !pieceStorage_->getDiskAdaptor()->fileAllocationIterator()->finished();
 }
 
@@ -1204,7 +1204,7 @@ void RequestGroup::reportDownloadFinished()
 #ifdef ENABLE_BITTORRENT
   if(downloadContext_->hasAttribute(bittorrent::BITTORRENT)) {
     TransferStat stat = calculateStat();
-    uint64_t completedLength = getCompletedLength();
+    off_t completedLength = getCompletedLength();
     double shareRatio = completedLength == 0 ? 0.0 :
       1.0*stat.getAllTimeUploadLength()/completedLength;
     SharedHandle<TorrentAttribute> attrs =
